included the weather service on the location_screen. Modified the tuype of the received parameter on the method getMessage of the WeatherModel, from int to double. Reduce the kTempTextStyle fontSize from 100 to 80. Added units in metric to the url of the method getDataByCoordinates at 'services/networking.dart:11'

+++ lib/services/weather.dart
@@ -19,7 +19,7 @@ class WeatherModel {
     }
   }
 
-  String getMessage(int temp) {
+  String getMessage(double temp) {
     if (temp > 25) {
       return 'It\'s 🍦 time';
     } else if (temp > 20) {
